# swagger.yaml
...
swagger: "2.0"
info:
  description: This is a API of SaratovNews
  version: 1.0.0
  title: Simple API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: virtserver.swaggerhub.com
basePath: /Ometh322/SararovNews/1.0.0
...
schemes:
- https
- http
paths:
...
  /events:
    get:
      tags:
      - events
      summary: List All Events
      description: Returns the whole Events
      operationId: listEvents
      produces:
      - application/json
      - application/xml
      parameters: []
      responses:
        "200":
          description: Successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Events'
        "404":
          description: Events not found
    put:
      tags:
      - events
      summary: Adds a new Events post
      operationId: addEvent
      consumes:
      - application/json
      - application/xml
      produces:
      - application/json
      - application/xml
      parameters:
      - in: body
        name: body
        description: Events object that needs to be added
        required: true
        schema:
          $ref: '#/definitions/Events'
      responses:
        "201":
          description: Successful creating
          schema:
            $ref: '#/definitions/Events'
        "405":
          description: Invalid Input
  /events/{id}:
    get:
      tags:
      - events
      summary: Find Event by ID
      description: Returns a singleEvent
      operationId: getEventById
      produces:
      - application/json
      - application/xml
      parameters:
      - name: id
        in: path
        description: ID of Event to return
        required: true
        type: integer
        format: int64
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/Events'
        "400":
          description: Invalid ID supplied
        "404":
          description: Event not found
    post:
      tags:
      - events
      summary: Update events by ID
      operationId: updateEvents
      consumes:
      - application/json
      - application/xml
      produces:
      - application/json
      - application/xml
      parameters:
      - name: id
        in: path
        description: ID of Events that need to be updated
        required: true
        type: integer
        format: int64
      - in: body
        name: body
        description: Events object that needs to be updated
        required: true
        schema:
          $ref: '#/definitions/Events'
      responses:
        "200":
          description: Successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Event not found
        "405":
          description: Validation exception
    delete:
      tags:
      - events
      summary: Deletes a Event post
      operationId: deleteEvent
      produces:
      - application/json
      - application/xml
      parameters:
      - name: id
        in: path
        description: Event ID to delete
        required: true
        type: integer
        format: int64
      responses:
        "200":
          description: Successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Event not found
...
definitions:
...
  Events:
    allOf:
    - $ref: '#/definitions/Record'
    - properties:
        srartDate:
          type: string
          format: date-time
...
  Comments:
    type: object
    properties:
      id:
        type: integer
        format: int64
      creatorId:
        type: integer
        format: int64
      message:
        type: string
      publicationDate:
        type: string
        format: date-time
      numberOfLikes:
        type: integer
        format: int32
  Record:
    type: object
    properties:
      id:
        type: integer
        format: int64
      creatorId:
        type: integer
        format: int64
      title:
        type: string
      publicationDate:
        type: string
        format: date-time
      description:
        type: string
      numberOfComments:
        type: integer
        format: int32
      numberOfLikes:
        type: integer
        format: int32
      comments:
        type: array
        items:
          $ref: '#/definitions/Comments'
      typeOfRecord:
        type: string
        enum:
        - NEWS
        - EVENT
        - POLL
...